feat(CommandBar): code separation and new actions
